requirements.txt
flask
binance-connector
pandas
numpy
python-dotenv
